Revert "move `class OFFSETS`"
This reverts commit 8b3681fd45681abe6e5324eb0638e63afd458a5e.
