+++ source/dllmain.h
@@ -12,15 +12,47 @@
 #define _DLLMAIN_H_
 
 #define WIN32_LEAN_AND_MEAN
-#include "offsets.hpp"
 
-#ifdef _MSC_VER
 #pragma warning( disable : 4099 )
 #pragma warning( disable : 4250 )
 #pragma warning( disable : 4100 )
 #pragma warning( disable : 4733 )
 #pragma warning( disable : 4996 )
-#endif
+
+enum class OFFSETS : unsigned int {
+    // gtarp_clientside.asi + ...
+    GTARP_LOCKCONNECTTOOTHERSERVER = 0x5134,
+    GTARP_INITTEXTURE_INITHOOK = 0x1C4ED,
+    GTARP_INITTEXTURE_FNC = 0x1BEC0,
+
+    GTARP_ARRAYSERVERLOGO = 0xD8B40,
+    GTARP_SERVERID = 0xD372C,
+    GTARP_X2PAYDAY = 0xCDB20,
+
+    GTARP_DRAWHUD = 0x1C8B0,
+    GTARP_DRAWSPEEDOMETER_DRAWRAMKA = 0x25DDA,
+
+    GTARP_WANTEDLEVEL = 0xD8B60,
+
+    GTARP_DISABLECHECKTXD = 0xF154,
+
+    // samp.dll + ...
+    SAMP_ENABLECLOCK = 0xA0D84,
+
+    SAMP_CREATEFONT = 0xC650A,
+
+    SAMP_RENDERPLAYERTAG_FORMAT = 0x74E35,
+    SAMP_RENDERPLAYERTAG_PUSHBUFF_1 = 0x74E3A,
+    SAMP_RENDERPLAYERTAG_PUSHBUFF_2 = 0x74E80,
+
+    SAMP_FORMATPATHSCREENSHOT = 0x61B52,
+    SAMP_FORMATFULLPATHSCREENSHOT = 0xC4573,
+
+    // gta_sa.exe
+    GTASA_DRAWWANTED = 0x58D9A0,
+    GTASA_DRAWHUD_CLOCK_SPRINTF = 0x58EBAF,
+    GTASA_GAMELOOP = 0x53BEE0
+};
 
 // stl
 #include <string>
@@ -31,8 +63,6 @@
 #include <ctime>
 // win-api
 #include <Windows.h>
-
-#pragma comment(lib, "Crypt32.lib")
 // Hook Library
 #include <polyhook2/Detour/x86Detour.hpp>
 #include <polyhook2/CapstoneDisassembler.hpp>
@@ -40,22 +70,25 @@
 #pragma comment(lib, "capstone.lib")
 #pragma comment(lib, "PolyHook_2.lib")
 // Plugin SDK
+#define GTASA
+#define GTAGAME_NAME = "San Andreas"
+#define GTAGAME_ABBR = "SA"
+#define GTAGAME_ABBRLOW = "sa"
+#define GTAGAME_PROTAGONISTNAME = "CJ"
+#define GTAGAME_CITYNAME = "San Andreas"
+#define _LA_SUPPORT
+#define _DX9_SDK_INSTALLED
+#define PLUGIN_SGV_10US
+
 #include "plugin.h"
 #include "CSprite2d.h"
-
-#pragma comment(lib, "plugin.lib")
 // DirectX SDK
 #include <d3d9.h>
 #include <d3dx9.h>
-
-#pragma comment(lib, "d3d9.lib")
-#pragma comment(lib, "d3dx9.lib")
-// OpenSSL
-#pragma comment(lib, "libcrypto.lib")
-#pragma comment(lib, "libssl.lib")
 // My library
 #include "CLog.hpp"
 extern CLog* g_pLog;
+//#include "TimeZone.hpp"
 
 #include "CConfig.hpp"
 #include "patch.hpp"
