--- source/offsets.hpp
@@ -1,41 +1,0 @@
-#ifndef PATCHGTARPCLIENT_OFFSETS_HPP
-#define PATCHGTARPCLIENT_OFFSETS_HPP
-
-enum class OFFSETS : unsigned int {
-    // gtarp_clientside.asi + ...
-    GTARP_LOCKCONNECTTOOTHERSERVER = 0x5134,
-    GTARP_INITTEXTURE_INITHOOK = 0x1C4ED,
-    GTARP_INITTEXTURE_FNC = 0x1BEC0,
-
-    GTARP_ARRAYSERVERLOGO = 0xD8B40,
-    GTARP_SERVERID = 0xD372C,
-    GTARP_X2PAYDAY = 0xCDB20,
-
-    GTARP_DRAWHUD = 0x1C8B0,
-    GTARP_DRAWSPEEDOMETER_DRAWRAMKA = 0x25DDA,
-
-    GTARP_WANTEDLEVEL = 0xD8B60,
-
-    GTARP_DISABLECHECKTXD = 0xF154,
-
-    // samp.dll + ...
-    SAMP_ENABLECLOCK = 0xA0D84,
-
-//    SAMP_CREATEFONT = 0xC650A,
-    SAMP_GETFONTSIZE = 0xC5B20,
-    SAMP_GETFONTWEIGHT = 0xC5BD0,
-
-    SAMP_RENDERPLAYERTAG_FORMAT = 0x74E35,
-    SAMP_RENDERPLAYERTAG_PUSHBUFF_1 = 0x74E3A,
-    SAMP_RENDERPLAYERTAG_PUSHBUFF_2 = 0x74E80,
-
-    SAMP_FORMATPATHSCREENSHOT = 0x61B52,
-    SAMP_FORMATFULLPATHSCREENSHOT = 0xC4573,
-
-    // gta_sa.exe
-    GTASA_DRAWWANTED = 0x58D9A0,
-    GTASA_DRAWHUD_CLOCK_SPRINTF = 0x58EBAF,
-    GTASA_GAMELOOP = 0x53BEE0
-};
-
-#endif //PATCHGTARPCLIENT_OFFSETS_HPP
